+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/ProjectCard.module.scss
